// src/app/types/MovieTypes.tsx
export type TMovie = {
  id: number;
  name_eng: string;
  name: string;
  subtitle?: string | null;
  year: string;
  country: string;
  imdb: string;
  creator: string;
  actors?: string | null;
  description: string;
  players: string;
  trailer?: string | null;
  poster_url?: string | null;
  thumbnail_url?: string | null;
  type: number;
  addons: string;
  genres?: string | null;
  views?: number | null;
  create_date: string;
};
export type TMovieCard = {
  id: number;
  name_eng: string;
  name: string;
  subtitle?: string | null;
  year: string;
  country: string;
  imdb: string;
  creator: string;
  actors?: string | null;
  description: string;
  players: string;
  trailer?: string | null;
  poster_url?: string | null;
  thumbnail_url?: string | null;
  type: number;
  addons: string;
  genres?: string | null;
  views?: number | null;
  create_date: string;
};